Extract ProfileHeaderButtonStyle to new file
